requirements.txt
gradio
rembg
Pillow>=9.2.0
flask